third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-014-expected.txt


FAIL CSS Test: @viewport constrained - zoom greater than max-zoom. assert_equals: expected 1600 but got 800
